+++ src/app/manifestGeneration/manifestGenerators/generator15/ManifestGenerator.ts
@@ -3,9 +3,9 @@
 // Licensed under the MIT license.
 
 import { BaseManifestGenerator } from "../../BaseManifestGenerator";
-import { TabManifestUpdater } from "./TabManifestUpdater";
-import { BotManifestUpdater } from "./BotManifestUpdater";
-import { ConnectorManifestUpdater } from "../generator13/ConnectorManifestUpdater"; // same as v1.3
+import { TabManifestUpdater } from "../generator14/TabManifestUpdater";
+import { BotManifestUpdater } from "../generator14/BotManifestUpdater";
+import { ConnectorManifestUpdater } from "../generator13/ConnectorManifestUpdater";
 import { MessageExtensionManifestUpdater } from "./MessageExtensionManifestUpdater";
 import { GeneratorTeamsAppOptions } from "../../../GeneratorTeamsAppOptions";
 import * as chalk from 'chalk';
@@ -21,26 +21,32 @@ export class ManifestGenerator extends BaseManifestGenerator {
 
     public generateManifest(options: GeneratorTeamsAppOptions): any {
         const manifest = super.generateManifest(options);
-        manifest["$schema"] = "https://developer.microsoft.com/en-us/json-schemas/teams/v1.4/MicrosoftTeams.schema.json";
-        manifest.manifestVersion = "1.4";
+        manifest["$schema"] = "https://developer.microsoft.com/en-us/json-schemas/teams/v1.5/MicrosoftTeams.schema.json";
+        manifest.manifestVersion = "1.5";
+
+        if (options.mpnId !== undefined) {
+            manifest.developer.mpnId = options.mpnId;
+        }
         return manifest;
     }
 
     public supportsUpdateManifest(from: string): boolean {
-        return from === "1.3";
+        return from === "1.3" || from === "1.4";
     }
 
     public updateManifest(manifest: any, log?: (message?: string, context?: any) => void): any {
-        if (manifest.manifestVersion === "1.3") {
-            manifest["$schema"] = "https://developer.microsoft.com/en-us/json-schemas/teams/v1.4/MicrosoftTeams.schema.json";
-            manifest.manifestVersion = "1.4";
+        if (manifest.manifestVersion === "1.3" || manifest.manifestVersion === "1.4") {
+            manifest["$schema"] = "https://developer.microsoft.com/en-us/json-schemas/teams/v1.5/MicrosoftTeams.schema.json";
+            manifest.manifestVersion = "1.5";
 
-            if(manifest.composeExtensions) {
+            if (manifest.composeExtensions) {
                 manifest.composeExtensions.forEach((composeExtension: { commands: { title: string; type: string; }[] }) => {
-                    if(composeExtension.commands) {
-                        composeExtension.commands.forEach( (command: { title: string; type: string; }) => {
-                            if(log) log(chalk.default.whiteBright(`Updating Message Extension "${command.title}" with the "type" property set to "query"`));
-                            command.type = "query";
+                    if (composeExtension.commands) {
+                        composeExtension.commands.forEach((command: { title: string; type: string; }) => {
+                            if (command.type === undefined) {
+                                if (log) log(chalk.default.whiteBright(`Updating Message Extension "${command.title}" with the "type" property set to "query"`));
+                                command.type = "query";
+                            }
                         });
                     }
                 });