+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*Unreleased]
+
+### Added
+
+* Added support for manifest version 1.5, new default
+* Added suppot for optionally adding Microsoft Partner Id. More info at https://aka.ms/partner 
+* Added files support option for bots (schema version 1.4+)
+
 ## [*2.8.0*] - <*2019-05-20*>
 
 ### Added